tests: refactoring and cleaning
