+++ topo_processor/file_system/tests/transfer_test.py
@@ -1,7 +1,4 @@
 import os
-import shutil
-from tempfile import mkdtemp
-from typing import Generator
 
 import pytest
 
@@ -9,19 +6,7 @@ from topo_processor.file_system.get_fs import get_fs
 from topo_processor.file_system.transfer import transfer_file
 
 
-@pytest.fixture(autouse=True)
-def setup() -> Generator[str, None, None]:
-    """
-    This function creates a temporary directory and deletes it after each test.
-    See following link for details:
-    https://docs.pytest.org/en/stable/fixture.html#yield-fixtures-recommended
-    """
-    target = mkdtemp()
-    yield target
-    shutil.rmtree(target)
-
-
-def test_transfer_local(setup) -> None:  # type: ignore
+def test_transfer_local(setup: str) -> None:
     dest_path = f"{setup}/test.tiff"
     input_path = os.path.join(os.getcwd(), "test_data/tiffs/SURVEY_1/CONTROL.tiff")
     transfer_file(input_path, "fakechecksum", "image/tiff", dest_path)
