+++ topo_processor/file_system/tests/write_json_test.py
@@ -1,26 +1,11 @@
 import os
-import shutil
-from tempfile import mkdtemp
-from typing import Generator
 
 import pytest
 
 from topo_processor.file_system.write_json import write_json
 
 
-@pytest.fixture(autouse=True)
-def setup() -> Generator[str, None, None]:
-    """
-    This function creates a temporary directory and deletes it after each test.
-    See following link for details:
-    https://docs.pytest.org/en/stable/fixture.html#yield-fixtures-recommended
-    """
-    target = mkdtemp()
-    yield target
-    shutil.rmtree(target)
-
-
-def test_write_json(setup) -> None:  # type: ignore
+def test_write_json(setup: str) -> None:
     my_dict = {"foo": "foo", "bar": 1}
     target = setup + "/test.json"
     write_json(my_dict, target)
